+++ sdks/python/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.1] - 2025-08-26
+
+### Added
+
+- Sessions can be deleted
+- Message metadata can be updated
+
 ## [1.4.0] - 2025-08-12
 
 ### Added